backend/requirements.txt
pandas
matplotlib
seaborn
openpyxl
fastapi
SQLAlchemy[asyncio]
pydantic
asyncpg
uvicorn[standard]
requests>=2.32.3
python-dotenv
dbt-core
dbt-postgres



